requirements.txt
pytest==5.2.1
xlrd==1.2.0
loguru==0.4.1
requests==2.23.0
selenium==3.141.0
PyMySQL==0.9.3
Appium_Python_Client==0.51
beautifulsoup4==4.9.0
Jinja2==2.11.2
paramiko==2.7.1
